+++ modules/iam-eks-role/main.tf
@@ -1,6 +1,9 @@
 data "aws_caller_identity" "current" {}
-
 data "aws_partition" "current" {}
+
+locals {
+  role_name_condition = var.role_name != null ? var.role_name : "${var.role_name_prefix}*"
+}
 
 data "aws_eks_cluster" "main" {
   for_each = var.cluster_service_accounts
@@ -10,11 +13,32 @@ data "aws_eks_cluster" "main" {
 
 data "aws_iam_policy_document" "assume_role_with_oidc" {
   dynamic "statement" {
+    # https://aws.amazon.com/blogs/security/announcing-an-update-to-iam-role-trust-policy-behavior/
+    for_each = var.allow_self_assume_role ? [1] : []
+
+    content {
+      sid     = "ExplicitSelfRoleAssumption"
+      effect  = "Allow"
+      actions = ["sts:AssumeRole"]
+
+      principals {
+        type        = "AWS"
+        identifiers = ["*"]
+      }
+
+      condition {
+        test     = "ArnLike"
+        variable = "aws:PrincipalArn"
+        values   = ["arn:${local.partition}:iam::${local.account_id}:role${var.role_path}${local.role_name_condition}"]
+      }
+    }
+  }
+
+  dynamic "statement" {
     for_each = var.cluster_service_accounts
 
     content {
-      effect = "Allow"
-
+      effect  = "Allow"
       actions = ["sts:AssumeRoleWithWebIdentity"]
 
       principals {