+++ modules/iam-eks-role/variables.tf
@@ -63,3 +63,9 @@ variable "max_session_duration" {
   type        = number
   default     = 43200
 }
+
+variable "allow_self_assume_role" {
+  description = "Determines whether to allow the role to be [assume itself](https://aws.amazon.com/blogs/security/announcing-an-update-to-iam-role-trust-policy-behavior/)"
+  type        = bool
+  default     = false
+}